% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiet_library.R
\name{quiet_library}
\alias{quiet_library}
\title{Quietly load package}
\usage{
quiet_library(pkgs)
}
\arguments{
\item{pkgs}{Character vector of packages.}
}
\description{
Utility function for suppressing as much console output as
possible when calling `library`, used mainly in scripts that
are meant to be run in batch with Rscript.
}
\seealso{
Other utility functions: \code{\link{whoami}}
}
