% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whoami.R
\name{whoami}
\alias{whoami}
\title{Get name the executing script file}
\usage{
whoami()
}
\value{
String of basename.
}
\description{
Utility function that, when used with Rscript, detects the file
running and returns a vector the basename without extension.
}
