% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr2tikz.R
\name{expr2tikz}
\alias{expr2tikz}
\title{Translate R expression into tikz-qtree and save figure}
\usage{
expr2tikz(expr, filename = NULL, path = getwd(), dpi = 600,
  keep_tex = FALSE)
}
\arguments{
\item{expr}{expression

  NB: the current version can't handle named arguments, neither in function calls nor definitions.}

\item{filename}{file name with extension (either .png, .svg, or .pdf) or NULL for not
saving an output file.}

\item{path}{subdirectory in which to save the figure}

\item{dpi}{image density, only relevant for png output}

\item{keep_tex}{should the .tex file be saved as well?

  This argument is ignored if filename is NULL.}
}
\value{
qtree string, invisibly.
  If a filename was provided, it will have node styles, otherwise it will be
  as plain as possible and printed to the console.
}
\description{
Plots the abstract syntax tree of an expression in a pdf, png, or svg file.
The qtree string assumes the node styles \code{call}, \code{sym}, and
\code{lit} to exist (call, symbol, and literal nodes, respectively).
See below for system requirements.
}
\details{
This function is heavily inspired by the abstract syntax tree figures in
Hadley Wickham's \href{https://adv-r.hadley.nz}{Advanced R},
Section 20 (Expressions), and the tikz styles are made to closely mimic the
style of those figures.
The code itself is adapted from the function \link[lobstr]{ast}.
}
\section{System requirements}{

In order to render the tree, the following components are required to be installed
on the system:

For pdf figures, \code{pdflatex} must be installed (e.g. with \href{https://www.tug.org/texlive/}{TeXLive}
or \href{https://miktex.org/}{MiKTeX}) and on the PATH, and the following packages must be installed:
\itemize{
  \item fontenc
  \item xcolor
  \item tikz
  \item tikz-qtree
  \item inconsolata
}

For png figures, the requirements for pdf figures must be met, and in addition, \code{convert} from the
\href{https://www.imagemagick.org/script/index.php}{ImageMagick} suite must be installed and on the PATH.

For svg figures, the requirements for pdf figures must be met, and in addition, \code{pdf2svg} must
be installed and on the PATH.
}

\examples{
expr2tikz(y <- 2 * x, "fig.png")

}
