% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whoami.R
\name{whoami}
\alias{whoami}
\title{Get name the executing script file}
\usage{
whoami()
}
\value{
String of basename without file extension.
}
\description{
Utility function that tries to detecs the file being run,
either with Rscript or with `source`, and returns the
basename without file extension.
If the function is not called as a result of a file being
run, it throws a warning and returns an empty string.
}
\seealso{
Other utility functions: \code{\link{quiet_library}}
}
